Restructing into packages and removing cyclic dependencies
