+++ Makefile
@@ -1,7 +1,10 @@
+NAME := ras-rm-printfile
+MAIN_SRC_FILE=cmd/ras-rm-print-file/main.go
+
 .PHONY: test
 test:
-	go test
+	go test  ./...
 
 .PHONY: build
 build:
-	go build
+	go build -o build/$(NAME) $(MAIN_SRC_FILE)
