+++ cmd/ras-rm-print-file/main_test.go
