+++ internal/sftp/sftp_upload_test.go
@@ -1,0 +1,1 @@
+package sftp
