--- main.go
@@ -1,130 +1,0 @@
-package main
-
-import (
-	"encoding/json"
-	"fmt"
-	"github.com/gorilla/mux"
-	log "github.com/sirupsen/logrus"
-	"github.com/spf13/viper"
-	"io/ioutil"
-	"net/http"
-)
-
-func middleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		log.Debug("setting content type to application/json")
-		w.Header().Add("Content-Type", "application/json")
-		next.ServeHTTP(w, r)
-	})
-}
-
-func print(w http.ResponseWriter, r *http.Request) {
-	switch r.Method {
-	case "POST":
-		log.Debug("post request received processing")
-		reqBody, err := ioutil.ReadAll(r.Body)
-		if err != nil {
-			log.WithError(err).Error("unable to read body")
-			w.WriteHeader(http.StatusInternalServerError)
-		}
-		log.WithField("reqBody", reqBody).Debug("body of request")
-		vars := mux.Vars(r)
-		filename := vars["filename"]
-		if filename == "" {
-			w.WriteHeader(http.StatusBadRequest)
-			fmt.Fprintln(w, "missing filename")
-		}
-		log.WithField("filename", filename).Info("received request to print file")
-		var printFileEntries []*PrintFileEntry
-		err = json.Unmarshal(reqBody, &printFileEntries)
-		if err != nil {
-			log.WithError(err).Error("unable to marshall json payload")
-			w.WriteHeader(http.StatusBadRequest)
-		}
-		printFile := PrintFile{
-			PrintFiles: printFileEntries,
-		}
-
-		w.WriteHeader(http.StatusAccepted)
-		resp, _ := json.Marshal(printFile)
-		log.WithField("resp", string(resp)).Debug("about to process")
-		//spawn a process to process the print file
-		go Process(filename, &printFile)
-	default:
-		w.WriteHeader(http.StatusMethodNotAllowed)
-		log.Info("print - method not allowed")
-		fmt.Fprintf(w, "Only POST methods are supported.")
-	}
-}
-
-func alive(w http.ResponseWriter, r *http.Request) {
-	switch r.Method {
-	case "GET":
-		log.Debug("alive OK")
-		w.WriteHeader(http.StatusOK)
-		fmt.Fprint(w, "{\"status\": \"OK\"}")
-	default:
-		log.Debug("alive -method not allowed")
-		w.WriteHeader(http.StatusMethodNotAllowed)
-		fmt.Fprint(w, "Only GET methods are supported")
-	}
-}
-
-func ready(w http.ResponseWriter, r *http.Request) {
-	switch r.Method {
-	case "GET":
-		log.Debug("ready OK")
-		w.WriteHeader(http.StatusOK)
-		fmt.Fprint(w, "{\"status\": \"READY\"}")
-	default:
-		log.Debug("ready - method not allowed")
-		w.WriteHeader(http.StatusMethodNotAllowed)
-		fmt.Fprint(w, "Only GET methods are supported")
-	}
-}
-
-func configureLogging() {
-	logLevel := viper.GetString("LOG_LEVEL")
-	level, err := log.ParseLevel(logLevel)
-	if err != nil {
-		log.WithError(err).WithField("logLevel", logLevel).Error("invalid log level")
-		//default to debug
-		level = log.DebugLevel
-	}
-	log.SetLevel(level)
-	log.WithField("level", logLevel).Debug("log level set")
-}
-
-func setDefaults() {
-	viper.SetDefault("LOG_LEVEL", "debug")
-	viper.SetDefault("BUCKET_NAME", "ras-rm-printfile")
-	viper.SetDefault("GOOGLE_CLOUD_PROJECT", "ras-rm-sandbox")
-	viper.SetDefault("SFTP_HOST", "localhost")
-	viper.SetDefault("SFTP_PORT", "22")
-	viper.SetDefault("SFTP_USERNAME", "sftp")
-	viper.SetDefault("SFTP_PASSWORD", "sftp")
-	viper.SetDefault("SFTP_DIRECTORY", "printfiles")
-}
-
-func configure() {
-	//config
-	viper.AutomaticEnv()
-	setDefaults()
-	configureLogging()
-}
-
-func main() {
-	configure()
-	log.Info("starting ras-rm-print-file")
-
-	//configure the gorilla router
-	r := mux.NewRouter()
-	r.Use(middleware)
-	r.HandleFunc("/print/{filename}", print)
-	r.HandleFunc("/alive", alive)
-	r.HandleFunc("/ready", ready)
-	http.Handle("/", r)
-
-	log.Info("started")
-	log.Fatal(http.ListenAndServe(":8080", nil))
-}
