--- printfile_test.go
@@ -1,2 +1,0 @@
-package main
-
